Added a form for domains.
If the domain obj_id is used in the URL, then the domain form is shown.
If the domain path is used in the URL, then a datatable with the list of domains below that path is shown instead.
There's still an issue when deleting a domain, because it will be kept in the sidebar domain list until a new domain is created.