Cleaned default conversation ids list and ignored empty mail.provider passwd.
The list of default conversations for a given domain is now cleaned before saving. This is needed, because webix, for some reason, pass a list of empty values mixed with real conversation ids.

Also ignored the password field when updating a mail.provider. This will allow the user to edit a mail provider without submitting the password/API key