Added a function to set a field value to a group of objects below a certain domain.
This function can be set to:
- Look for objects of a certain type
- Look for objects in depth (grandchildren of the specified domain)
- Use a fun to retrieve the property value from the object map
- Use a fun to store the new value in the object map

Prior to all operations, the specified domain and all its children will be unloaded from memory and the domain will be disabled to prevent them from loading during the procedure. When there isn't any object left, the specified domain will be re-enabled.

The function fix_push_srv_id was added as an example. This function looks for conversations below a certain domain and updates the "push_srv_id" to a certain common value (in case it was needed) writing directly to ES.