Added delete_missing_user_content.
This function parses all objects from a given domain and checks if some of them should be deleted/removed:
- Messages written by deleted users
- Messages quoted from deleted users
- Messages from deleted conversations
- Deleted members from a conversation
- Direct conversations with deleted members
- Conversations created by deleted users

Only messages and conversations are checked.
If called with just the domain, then debug mode will be enabled and it will only list what actions would be performed.
To effectively delete those objects, the second boolean value must be set to "true".
If some objects have been deleted, then the domain will be unloaded to force a refresh for users still logged in.