Deactivate push devices when a session is stopped.
When the session object is stopped, the session is unregistered from the user object.
At that moment, the push device is deactivated by searching for the session_id attached to said device.
The active flag and the session_id will be kept in memory while the user is loaded, but neither of them will be persisted to ES.