Added filter_obj_id/1 and filter_obj_ids/1.
These utility functions parse an obj_id or an array of obj_ids and remove the ones that point to a non-existent obj_id.
By using this function instead of parse_obj_ids in the object_parse callback, an object can be loaded even though the object it points to was deleted while the former was not loaded into memory.